Result of running  python-modernize -w pyth6
modernize automates most of the changes required to make
Python 2 code compatible with Python 3.
The resulting code will rely on one additional package: six

Some more changes will be required for code such as this one
that handes both binary data and text data without always being
explicit in the original Python 2 code which is which.

The file  modernize_output.txt  captures the diff of these changes
as shown by the modernize call.
The file  modernize_output_strippeddown.txt is a subset of that
previous file (as described in the header of itself).