I have repaired many of the bytes-vs-str issues in pyth\plugins\rtf15\reader.py pyth\plugins\xhtml\writer.py
The former in particular was tricky because most strings have to be handled
as bytestrings -- but not all of them.
See  http://pythonhosted.org/six/

These two now appear to work for ASCII and 8-bit non-ASCII characters in the RTF file,
at least for a simple RTF file.
Complex files and true Unicode remain to be seen.