established end-to-end tests for Rtf15Reader
They are based on pairs of files with the same basename in directories
tests/rtfs           (inputfiles) and
tests/rtf-as-html    (reference outputfiles)
The unittest test method for each pair is created dynamically
by test_readrtf15.py and will write the actual output to
yet another file with that same basename in directory
tests/currentoutput  (files holding actual outputs).
Those files are deleted only if the test succeeds and so
can be used for analysis if the test fails.