completely reorganized the RTF-to-XHTML-based tests
I have thrown out the complex test files (except the interesting zh-cn.rtf)
as well as the overly simple or redundant test files and have introduced
a set of simple tests (not covering too much functionality at once)
and a regular naming of the test files.
The file names in tests/rtfs now consist of two parts:
program-testcontent.rtf
where program is one of
- msword: Microsoft Word from Microsoft Office 2013 on Windows 7
- librewriter: Writer from LibreOffice 4.4 on Windows 7
- wordpad: Microsoft Wordpad on Windows 7
and testcontent is a name describing roughly what functionality
is tested in the file.

The corresponding found-to-be-correct test outputs are in
tests/rtf-as-html/*.html
where these outputs have been truncated to zero bytes
for test cases with errors in their output.